feat: Separate interface files
