+++ src/app/interfaces/price.interface.ts
@@ -1,0 +1,16 @@
+
+export interface Product {
+  name: string;
+  price: number;
+  date: string;
+  companySign: string;
+  companyName: string;
+}
+
+export interface ProductApiResponse {
+  Name: string;
+  Price: number;
+  Date: string;
+  company_sign: string;
+  company_name: string;
+}
